+++ apk/Changelog.md
@@ -9,12 +9,18 @@ The App is divided in two part, **Apk** and **Web** you can check it version in 
 
 The app has a building update check function, so whenever there is a update a red text will show up bellow the main pages clock, also a short warning will pop on the screen let you know if is a **Apk** or **Web** update.
 
+Apk Version 3.0.268 - Web Version October 22 2020
+==============
+
+* Automatic cleanup history unreachable Lives that became a VOD but the VOD is now deleted
+* General performance improves and bug fixes
+
 Web Version October 21 2020
 ==============
 
 * Add a progress indicator to already watched VOD/Clip
 * Save a clip as watched when the previewed clip ends (of course if history is enabled for clips)
-* Automatic remove from history unreachable (deleted from the server) VOD/Clip
+* Automatic cleanup history unreachable (deleted from the server) VOD/Clip
 * General performance improves and bug fixes
 
 Web Version October 20 2020